requirements.txt
django
djangorestframework
langchain
langgraph
openai
chromadb
python-dotenv
datasets
langchain-community
tiktoken
pymupdf
pypdf